# backend/docs/swagger.yaml
basePath: /api
definitions:
  github_com_xilo-platform_xilo_internal_auth_model.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.User'
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.RequestOTPRequest:
    properties:
      phone:
        type: string
      purpose:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.VerifyOTPLoginRequest:
    properties:
      code:
        type: string
      phone:
        type: string
    type: object
  github_com_xilo-platform_xilo_internal_auth_model.VerifyOTPRegisterRequest:
    properties:
      code:
        type: string
      email:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@xilo.example.com
    name: Xilo API Support
    url: https://xilo.example.com/support
  description: A self-hosted, Telegram-inspired modern blog platform API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://xilo.example.com/terms
  title: Xilo Blog Platform API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout current user and invalidate refresh token
      parameters:
      - description: Refresh token (optional)
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout current user
      tags:
      - auth
  /auth/me:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Update the profile of the currently authenticated user
      parameters:
      - description: Profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - auth
  /auth/otp/request:
    post:
      consumes:
      - application/json
      description: Request a one-time password for authentication
      parameters:
      - description: OTP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.RequestOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request OTP code
      tags:
      - auth
  /auth/otp/verify-login:
    post:
      consumes:
      - application/json
      description: Verify OTP code for login
      parameters:
      - description: OTP verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.VerifyOTPLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP for login
      tags:
      - auth
  /auth/otp/verify-register:
    post:
      consumes:
      - application/json
      description: Verify OTP code for registration
      parameters:
      - description: OTP registration verification
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.VerifyOTPRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify OTP for registration
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register with email, username, and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_xilo-platform_xilo_internal_auth_model.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"